% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytobankAPIstatsGUI.R
\name{CytobankAPIstatsGUI}
\alias{CytobankAPIstatsGUI}
\title{Exports processed events and signaling data}
\usage{
CytobankAPIstatsGUI()
}
\description{
Exports processed events and signaling data
}
\examples{
 \dontrun{
library(CytobankAPIstats)
CytobankAPIstatsGUI()}
}
