% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfcsfiles.R
\name{getfcsfiles}
\alias{getfcsfiles}
\title{Gets fcs ID numbers and sample names from a given experiment}
\usage{
getfcsfiles(exptno, cyto_session)
}
\arguments{
\item{exptno}{- Integer representing an experiment ID on Cytobank account}

\item{cyto_session}{- API authentication token for session}
}
\value{
Returns a list of fcs file IDs with names of fcs files as names of list
}
\description{
Gets fcs ID numbers and sample names from a given experiment
}
\examples{
\donttest{
library(CytobankAPI)
cyto_session <- authenticate(site="premium", username="myusername", password="mypassword")
exptno<-2
getfcsfiles(exptno,cyto_session)}
}
