% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentevent.R
\name{percentevent}
\alias{percentevent}
\title{Calculates the percentage of cell populations given an experiment}
\usage{
percentevent(cyto_session, markersofinterest, popsofinterest, exptID, grouping,
  specimennames, means)
}
\arguments{
\item{cyto_session}{- API authentication token for session}

\item{markersofinterest}{- List of names of channel parameters in Cytobank}

\item{popsofinterest}{- List of populations of interest to calculate percentages with reference population for percentages listed first}

\item{exptID}{- Integer representing an experiment ID on Cytobank account}

\item{grouping}{- A list of indices corresponding to samples from the same donor ex list(c(1,2),c(3,4,5)) if rows 1 and 2 are from pt1,3,4,5 are from pt2, etc.}

\item{specimennames}{- List of specimen names as strings; needs to be same length as number of groupings}

\item{means}{- A boolean if mean =TRUE, a mean for all groups in the variable group  is calculated, otherwise individual means are returned.}
}
\value{
- Returns either the percentage or mean percentage per specimen of each cell type, as specified by mean parameter
}
\description{
Calculates the percentage of cell populations given an experiment
}
\examples{
\donttest{
library(CytobankAPI)
cyto_session <- authenticate(site="premium", username="myusername", password="mypassword")
markersofinterest<-c("CD3","CD56")
popsofinterest<-c("CD4 T cells","NK cells")
exptID=4
grouping<-list(c(1,2),c(3,4,5),c(6,7))
specimennames<-c("Patient1","Patient2","Control1")
means=T
percentevent(cyto_session,markersofinterest,popsofinterest,exptID,grouping,specimennames,means)
}
}
